+++ src/main/java/com/qa/choonz/service/ArtistService.java
@@ -1,56 +1,66 @@
 package com.qa.choonz.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.stereotype.Service;
-
 import com.qa.choonz.exception.ArtistNotFoundException;
 import com.qa.choonz.persistence.domain.Artist;
+import com.qa.choonz.persistence.domain.UserDetails;
 import com.qa.choonz.persistence.repository.ArtistRepository;
+import com.qa.choonz.persistence.roles.UserRole;
 import com.qa.choonz.rest.dto.ArtistDTO;
+import com.qa.choonz.rest.mapper.ArtistMapper;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class ArtistService {
 
     private ArtistRepository repo;
-    private ModelMapper mapper;
+    private ArtistMapper mapper;
 
-    public ArtistService(ArtistRepository repo, ModelMapper mapper) {
+    public ArtistService(ArtistRepository repo, ArtistMapper mapper) {
         super();
         this.repo = repo;
         this.mapper = mapper;
     }
 
-    private ArtistDTO mapToDTO(Artist artist) {
-        return this.mapper.map(artist, ArtistDTO.class);
-    }
+    public ArtistDTO create(Artist artist, UserDetails user) {
+        // ensure only admin can create artists
+        if (user == null || user.getRole() != UserRole.ADMIN){return null;}
 
-    public ArtistDTO create(Artist artist) {
         Artist created = this.repo.save(artist);
-        return this.mapToDTO(created);
+        return mapper.mapToDeepDTO(created);
     }
 
     public List<ArtistDTO> read() {
-        return this.repo.findAll().stream().map(this::mapToDTO).collect(Collectors.toList());
+        return this.repo.findAll().stream().map(mapper::mapToDeepDTO).collect(Collectors.toList());
     }
 
     public ArtistDTO read(long id) {
         Artist found = this.repo.findById(id).orElseThrow(ArtistNotFoundException::new);
-        return this.mapToDTO(found);
+        return mapper.mapToDeepDTO(found);
     }
 
-    public ArtistDTO update(Artist artist, long id) {
+    public ArtistDTO update(Artist artist, long id, UserDetails user) {
+        // ensure only admin can edit artists
+        if (user == null || user.getRole() != UserRole.ADMIN){return null;}
+
         Artist toUpdate = this.repo.findById(id).orElseThrow(ArtistNotFoundException::new);
         toUpdate.setName(artist.getName());
-        toUpdate.setAlbums(artist.getAlbums());
         Artist updated = this.repo.save(toUpdate);
-        return this.mapToDTO(updated);
+        return mapper.mapToDeepDTO(updated);
     }
 
-    public boolean delete(long id) {
+    public boolean delete(long id, UserDetails user) {
+        // ensure only admin can delete artists
+        if (user == null || user.getRole() != UserRole.ADMIN){return false;}
+
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
+    }
+
+    public Set<Artist> read(String term) {
+        return Set.copyOf(repo.findAllByNameContaining(term));
     }
 }
