+++ src/main/java/com/qa/choonz/service/TrackService.java
@@ -1,60 +1,73 @@
 package com.qa.choonz.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.stereotype.Service;
-
 import com.qa.choonz.exception.TrackNotFoundException;
 import com.qa.choonz.persistence.domain.Track;
+import com.qa.choonz.persistence.domain.UserDetails;
 import com.qa.choonz.persistence.repository.TrackRepository;
+import com.qa.choonz.persistence.roles.UserRole;
 import com.qa.choonz.rest.dto.TrackDTO;
+import com.qa.choonz.rest.mapper.TrackMapper;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class TrackService {
 
     private TrackRepository repo;
-    private ModelMapper mapper;
+    private TrackMapper mapper;
 
-    public TrackService(TrackRepository repo, ModelMapper mapper) {
+    public TrackService(TrackRepository repo, TrackMapper mapper) {
         super();
         this.repo = repo;
         this.mapper = mapper;
     }
 
-    private TrackDTO mapToDTO(Track track) {
-        return this.mapper.map(track, TrackDTO.class);
-    }
-
-    public TrackDTO create(Track track) {
+    public TrackDTO create(Track track, UserDetails user) {
+        // ensure only admin can create tracks
+        if (user == null || user.getRole() != UserRole.ADMIN) {
+            return null;
+        }
         Track created = this.repo.save(track);
-        return this.mapToDTO(created);
+        return mapper.mapToDeepDTO(created);
     }
 
     public List<TrackDTO> read() {
-        return this.repo.findAll().stream().map(this::mapToDTO).collect(Collectors.toList());
+        return this.repo.findAll().stream().map(mapper::mapToDeepDTO).collect(Collectors.toList());
     }
 
     public TrackDTO read(long id) {
         Track found = this.repo.findById(id).orElseThrow(TrackNotFoundException::new);
-        return this.mapToDTO(found);
+        return mapper.mapToDeepDTO(found);
     }
 
-    public TrackDTO update(Track track, long id) {
+    public TrackDTO update(Track track, long id, UserDetails user) {
+        // ensure only admin can edit tracks
+        if (user == null || user.getRole() != UserRole.ADMIN) {
+            return null;
+        }
+
         Track toUpdate = this.repo.findById(id).orElseThrow(TrackNotFoundException::new);
         toUpdate.setName(track.getName());
         toUpdate.setAlbum(track.getAlbum());
         toUpdate.setDuration(track.getDuration());
         toUpdate.setLyrics(track.getLyrics());
-        toUpdate.setPlaylist(track.getPlaylist());
-        Track updated = this.repo.save(toUpdate);
-        return this.mapToDTO(updated);
+        this.repo.save(toUpdate);
+        return mapper.mapToDeepDTO(this.repo.findById(id).orElseThrow(TrackNotFoundException::new));
     }
 
-    public boolean delete(long id) {
+    public boolean delete(long id, UserDetails user) {
+        // ensure only admin can delete tracks
+        if (user == null || user.getRole() != UserRole.ADMIN) {
+            return false;
+        }
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
     }
 
+    public Set<Track> read(String term) {
+        return Set.copyOf(repo.findAllByNameContaining(term));
+    }
 }
